+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix wrong probabilities for generating data in `cpm.applications.decision_making.PTSM2025` model [#67](https://github.com/DevComPsy/cpm/issues/67) @FloorJBurghoorn
 - Fix matplotlib>=3.10 dependency mismatch errors upon loading `cpm` by removing unused imports in `cpm.utils.metad`
 - Fix the usage of ppt_identifier in Minimize class ([#72](https://github.com/DevComPsy/cpm/pull/72)) @tzukpolinsky
+- Fix usage of NaN in `cpm.models.decision.SoftMax`
 
 ### Removed
 
@@ -53,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix `simulation_export` function to handle DataFrame output correctly
 - Fix `detailed_pandas_compiler` function to support new numpy versions
 - Fix probability adjustements in `cpm.optimisation.minimise.LogLikelihood` method to ensure correct parameter estimates
+- Fix NaN handling in `cpm.models.decision.Softmax`, and `cpm.models.decision.Sigmoid` due to infinities in the exponential function for out-of-bounds parameters
 
 ### Changed